Split lib/src/widgets/solid_security_key_manager_dialogs.dart into 2 files
